requirements-dev.txt
-r requirements.txt
ipython==9.4.0
jedi==0.19.2
Pygments==2.19.2
prompt_toolkit==3.0.51
traitlets==5.14.3
pexpect==4.9.0
ptyprocess==0.7.0
matplotlib-inline==0.1.7
asttokens==3.0.0
parso==0.8.4
pure_eval==0.2.3
executing==2.2.0
stack-data==0.6.3
ipython_pygments_lexers==1.1.1
decorator==5.2.1
wcwidth==0.2.13
django-browser-reload==1.18.0